Refactor: createDesign 폴더에 있는 컴포넌트 css 파일로 분리
