+++ src/pages/CreateDesign/Review/Review.css.ts
@@ -1,0 +1,20 @@
+import { Grid, Typography } from '@material-ui/core';
+import styled from 'styled-components';
+
+export const Title = styled(Typography)`
+  font-weight: 600;
+`;
+
+export const Row = styled(Grid)`
+  padding: 12px;
+`;
+
+export const Label = styled(Typography)`
+  padding-bottom: 10px;
+  width: 100%;
+`;
+
+export const Contents = styled(Typography)`
+  white-space: pre-line;
+  padding: 5px;
+`;
