+++ src/pages/CreateDesign/components/DesignSizeImage/DesignSizeImage.css.ts
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+
+export const DesignImageWrapper = styled.div`
+  display: flex;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+
+  img {
+    max-width: 100%;
+    max-height: 450px;
+    margin-top: 8px;
+  }
+`;
