+++ src/pages/CreateDesign/components/DesignSizeImage/index.tsx
@@ -1,19 +1,7 @@
 import KnitDesign from 'assets/designs/knit.png';
 import React from 'react';
-import styled from 'styled-components';
 
-const DesignImageWrapper = styled.div`
-  display: flex;
-  height: 100%;
-  align-items: center;
-  justify-content: center;
-
-  img {
-    max-width: 100%;
-    max-height: 450px;
-    margin-top: 8px;
-  }
-`;
+import { DesignImageWrapper } from './DesignSizeImage.css';
 
 const DesignSizeImage = (): React.ReactElement => {
   return (
