+++ src/pages/CreateDesign/components/Header/Header.css.ts
@@ -1,0 +1,12 @@
+import { Typography } from '@material-ui/core';
+import styled from 'styled-components';
+import { theme } from 'themes';
+
+export const Title = styled(Typography)`
+  margin-bottom: ${theme.spacing(2)};
+  font-weight: 600;
+`;
+
+export const Contents = styled(Typography)`
+  white-space: pre-line;
+`;
